% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_calculator.R
\name{bmi_calculator}
\alias{bmi_calculator}
\title{Calculate Body Mass Index (BMI)}
\usage{
bmi_calculator(weight, height)
}
\arguments{
\item{weight}{A numeric value representing weight in kilograms.}

\item{height}{A numeric value representing height in meters.}
}
\value{
The BMI value.
}
\description{
This function calculates the Body Mass Index (BMI) based on weight and height.
}
\examples{
bmi_calculator(70, 1.75)
}
