% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_calories.R
\name{daily_calories}
\alias{daily_calories}
\title{Calculate Daily Calorie Needs}
\usage{
daily_calories(weight, height, age, sex, activity_level)
}
\arguments{
\item{weight}{A numeric value representing weight in kilograms.}

\item{height}{A numeric value representing height in centimeters.}

\item{age}{A numeric value representing age in years.}

\item{sex}{A character value, either "male" or "female".}

\item{activity_level}{A character value representing activity level: "sedentary", "light", "moderate", "active", "very_active".}
}
\value{
The estimated daily calorie needs.
}
\description{
This function estimates the daily calorie needs based on activity level and personal data.
}
\examples{
daily_calories(70, 175, 30, "male", "moderate")
}
